+++ bungeecord-mysql/src/main/java/de/simonsator/partyandfriends/spigot/proxy/bungee/PartyAPIForSpigotBungeeBridgePlugin.java
