+++ party-spigot-api-for-party-and-friends/src/main/java/de/simonsator/partyandfriends/spigot/api/party/PlayerParty.java
