+++ party-spigot-api-for-party-and-friends/src/main/resources/plugin.yml
