+++ proxy-common/src/main/java/de/simonsator/partyandfriends/spigot/proxy/PartyBridgeProxyMySQLConnection.java
