+++ spigot-plugin/src/main/resources/config.yml
