+++ velocity-mysql/src/main/java/de/simonsator/partyandfriends/velocity/spigot/proxy/PartyAPIForSpigotVelocityBridgePlugin.java
